refactor: Move TransferOptimizedFilterFactory to model.costfilter package
